+++ apps/client/src/pages/interviews/index.tsx
@@ -4,115 +4,26 @@ import {
   GetServerSidePropsResult,
   InferGetServerSidePropsType,
 } from "next";
-import { useRouter } from "next/router";
 import { Button } from "@kokomen/ui/components/button";
-import { JSX, useState, useCallback, useMemo } from "react";
-import { useToast } from "@kokomen/ui/hooks/useToast";
-import { useMutation } from "@tanstack/react-query";
-import axios from "axios";
+import { JSX, useState } from "react";
 import Head from "next/head";
 import Image from "next/image";
 import Header from "@/shared/header";
-import {
-  NewInterviewRequest,
-  startNewInterview,
-} from "@/domains/interview/api";
+
 import { withCheckInServer } from "@/utils/auth";
-import {
-  Keyboard,
-  MicVocal,
-  Trophy,
-  Coins,
-  User as UserIcon,
-  Star,
-  Zap,
-} from "lucide-react";
+import { Trophy, Coins, User as UserIcon, Star, Zap } from "lucide-react";
 import { getUserInfo } from "@/domains/auth/api";
 import { User as UserType } from "@/domains/auth/types";
+import CreateInterviewForm from "@/domains/interview/components/createInterviewForm";
 
-const QUESTION_COUNT_OPTIONS = [5, 10, 15, 20] as const;
-
-type InterviewType = "text" | "voice";
-
-export default function Home({
+export default function InterviewMainPage({
   categories,
   userInfo,
 }: InferGetServerSidePropsType<typeof getServerSideProps>): JSX.Element {
-  const router = useRouter();
-  const { error: errorToast } = useToast();
-  const [interviewConfig, setInterviewConfig] = useState<NewInterviewRequest>({
-    category: categories[0].key || "",
-    max_question_count: QUESTION_COUNT_OPTIONS[0],
-  });
-  const [selectedInterviewType, setSelectedInterviewType] =
-    useState<InterviewType>("text");
+  const [selectedCategory, setSelectedCategory] = useState<Category>(
+    categories[0]
+  );
 
-  const createInterviewMutation = useMutation({
-    mutationFn: startNewInterview,
-    onSuccess: (data) => {
-      router.push({
-        pathname: `/interviews/${data.interview_id}`,
-      });
-    },
-
-    onError: (error) => {
-      if (axios.isAxiosError(error)) {
-        if (error.response?.status === 401) {
-          router.replace("/login");
-        }
-        errorToast({
-          title: "면접 생성 실패",
-          description: error.response?.data.message,
-        });
-      }
-    },
-    onSettled: () => {
-      console.log("Interview creation attempt completed");
-    },
-    retry: (failureCount, error) => {
-      if (axios.isAxiosError(error)) {
-        if (
-          error.response?.status &&
-          error.response.status >= 400 &&
-          error.response.status < 500
-        ) {
-          return false;
-        }
-
-        return failureCount < 2;
-      }
-      return false;
-    },
-    retryDelay: (attemptIndex) => Math.min(1000 * 2 ** attemptIndex, 30000),
-  });
-
-  const handleCategoryChange = useCallback((category: Category) => {
-    setInterviewConfig((prev) => ({ ...prev, category: category.key }));
-  }, []);
-
-  const handleQuestionCountChange = useCallback((event: "plus" | "minus") => {
-    setInterviewConfig((prev) => ({
-      ...prev,
-      max_question_count:
-        event === "plus"
-          ? Math.min(20, prev.max_question_count + 1)
-          : Math.max(3, prev.max_question_count - 1),
-    }));
-  }, []);
-
-  const handleInterviewTypeChange = useCallback((type: InterviewType) => {
-    setSelectedInterviewType(type);
-  }, []);
-
-  const handleNewInterview = useCallback(() => {
-    createInterviewMutation.mutate(interviewConfig);
-  }, [interviewConfig, createInterviewMutation]);
-
-  const selectedCategory = useMemo(() => {
-    return categories.find(
-      (c) => c.key === interviewConfig.category
-    ) as Category;
-  }, [categories, interviewConfig.category]);
   return (
     <>
       <Head>
@@ -137,14 +48,13 @@ export default function Home({
                       key={category.key}
                       role="tab"
                       className="text-sm font-semibold"
-                      aria-selected={interviewConfig.category === category.key}
+                      aria-selected={selectedCategory.key === category.key}
                       variant={
-                        interviewConfig.category === category.key
+                        selectedCategory.key === category.key
                           ? "primary"
                           : "text"
                       }
-                      onClick={() => handleCategoryChange(category)}
-                      disabled={createInterviewMutation.isPending}
+                      onClick={() => setSelectedCategory(category)}
                     >
                       {category.title}
                     </Button>
@@ -176,97 +86,7 @@ export default function Home({
                   </p>
                 </div>
 
-                {/* 설정 섹션 */}
-                <div className="grid md:grid-cols-2 gap-8 mb-12">
-                  {/* 문제 개수 선택 */}
-                  <div className="bg-fill-quaternary rounded-2xl p-6 border border-border">
-                    <h3 className="text-lg font-semibold text-text-heading mb-4 flex items-center gap-2">
-                      <div className="w-2 h-2 bg-primary rounded-full"></div>
-                      면접 문제 개수
-                    </h3>
-                    <div className="flex items-center justify-center gap-6">
-                      <Button
-                        onClick={() => handleQuestionCountChange("minus")}
-                        className="w-12 h-12 bg-bg-elevated rounded-full shadow-lg hover:shadow-xl transition-all duration-300 flex items-center justify-center text-2xl font-bold text-text-secondary hover:text-primary hover:scale-110 border border-border"
-                        variant="text"
-                      >
-                        -
-                      </Button>
-                      <div className="w-24 h-24 bg-primary rounded-full flex items-center justify-center shadow-xl">
-                        <span className="text-3xl font-bold text-text-light-solid">
-                          {interviewConfig.max_question_count}
-                        </span>
-                      </div>
-                      <Button
-                        onClick={() => handleQuestionCountChange("plus")}
-                        className="w-12 h-12 bg-bg-elevated rounded-full shadow-lg hover:shadow-xl transition-all duration-300 flex items-center justify-center text-2xl font-bold text-text-secondary hover:text-primary hover:scale-110 border border-border"
-                        variant="text"
-                      >
-                        +
-                      </Button>
-                    </div>
-                  </div>
-
-                  {/* 면접 타입 선택 */}
-                  <div className="bg-fill-quaternary rounded-2xl p-6 border border-border">
-                    <h3 className="text-lg font-semibold text-text-heading mb-4 flex items-center gap-2">
-                      <div className="w-2 h-2 bg-primary rounded-full"></div>
-                      면접 방식
-                    </h3>
-                    <div className="grid grid-cols-2 gap-3">
-                      <Button
-                        onClick={() => handleInterviewTypeChange("text")}
-                        className="py-6"
-                        variant={
-                          selectedInterviewType === "text"
-                            ? "primary"
-                            : "outline"
-                        }
-                      >
-                        <div className="flex flex-col items-center gap-2">
-                          <Keyboard className="w-6 h-6" />
-                          <span className="text-sm font-medium">텍스트</span>
-                        </div>
-                      </Button>
-                      <Button
-                        onClick={() => handleInterviewTypeChange("voice")}
-                        className="py-6"
-                        disabled
-                        variant={
-                          selectedInterviewType === "voice"
-                            ? "primary"
-                            : "outline"
-                        }
-                      >
-                        <div className="flex flex-col items-center gap-2">
-                          <MicVocal className="w-6 h-6" />
-                          <span className="text-sm font-medium">음성</span>
-                        </div>
-                      </Button>
-                    </div>
-                  </div>
-                </div>
-
-                {/* 면접 시작 버튼 */}
-                <div className="text-center flex flex-col items-center">
-                  <Button
-                    type="button"
-                    onClick={handleNewInterview}
-                    disabled={createInterviewMutation.isPending}
-                    size={"large"}
-                    className="font-semibold text-lg w-full"
-                  >
-                    {createInterviewMutation.isPending ? (
-                      <span>면접 시작 중...</span>
-                    ) : (
-                      <span>{interviewConfig.category} 면접 시작하기</span>
-                    )}
-                  </Button>
-                  <p className="mt-4 text-sm text-text-description">
-                    선택한 카테고리의 {interviewConfig.max_question_count}개
-                    문제로 면접을 진행합니다
-                  </p>
-                </div>
+                <CreateInterviewForm selectedCategory={selectedCategory} />
               </div>
             </div>
           </section>
